# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: RideMatch API
  version: 0.4.5
  description: |
    Contract for the RideMatch monolith (FastAPI). Source of truth for backend, frontend and tests.
    See CONTRACT.md for conventions, state machines, business rules, error codes and the WebSocket protocol.
    Authentication is handled by Clerk (sign-up, sign-in, password, email verification, social login).
    The backend never sees passwords; it verifies the Clerk session token on every request.
    Owned by @backend. Any change: bump `info.version`, commit, then message @frontend and @tests.
servers:
  - url: http://localhost:8000/api/v1
    description: Local dev
  - url: https://api.example.com/api/v1
    description: Production (placeholder domain)
security:
  - bearerAuth: []
...
paths:
...
  /rides:
    post:
      tags: [rides]
      operationId: createRide
      summary: |
        Driver offers a ride. status=upcoming, available_seats=capacity.
        The caller must have a vehicle on their profile (409 VEHICLE_REQUIRED).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RideCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ride" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }         # VEHICLE_REQUIRED
        "422": { $ref: "#/components/responses/ValidationError" }  # incl. DEPARTURE_IN_PAST
  /rides/mine:
    get:
      tags: [rides]
      operationId: listMyRides
      summary: Rides the caller offers as driver, ordered by departure_time ascending
      parameters:
        - name: status
          in: query
          description: Comma-separated filter, e.g. `upcoming,full`
          schema: { type: string }
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Offset" }
      responses:
        "200":
          description: Rides
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Ride" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationError" }  # unknown status filter
  /rides/{ride_id}:
    get:
      tags: [rides]
      operationId: getRide
      parameters: [{ $ref: "#/components/parameters/RideId" }]
      responses:
        "200":
          description: Ride
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ride" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [rides]
      operationId: updateRide
      summary: |
        Driver only, status must be `upcoming` or `full`. If the ride has approved passengers,
        location/departure_time changes → 409 RIDE_HAS_APPROVED_PASSENGERS, and capacity may not drop
        below approved seats → 409 CAPACITY_BELOW_APPROVED. Changing capacity recomputes available_seats
        and status (full ↔ upcoming). Sending a locked field counts as a change even if the value is
        unchanged. A new departure_time must still be in the future (422 DEPARTURE_IN_PAST).
      parameters: [{ $ref: "#/components/parameters/RideId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RideUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ride" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /rides/{ride_id}/cancel:
    post:
      tags: [rides]
      operationId: cancelRide
      summary: |
        Driver only, from `upcoming` or `full`. Sets status=cancelled, all pending/approved requests →
        cancelled, notifies approved passengers (ride_cancelled) and pending requesters. No approved
        request is left, so available_seats goes back to capacity.
      parameters: [{ $ref: "#/components/parameters/RideId" }]
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ride" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }  # INVALID_STATE_TRANSITION
  /rides/{ride_id}/start:
    post:
      tags: [rides]
      operationId: startRide
      summary: Driver only, from `upcoming`/`full`, on departure day. Pending requests → rejected.
      parameters: [{ $ref: "#/components/parameters/RideId" }]
      responses:
        "200":
          description: Started (status=in_progress)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ride" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }  # INVALID_STATE_TRANSITION, TOO_EARLY_TO_START
  /rides/{ride_id}/complete:
    post:
      tags: [rides]
      operationId: completeRide
      summary: Driver only, from `in_progress`. Opens rating for all participants.
      parameters: [{ $ref: "#/components/parameters/RideId" }]
      responses:
        "200":
          description: Completed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ride" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Clerk session token (short-lived, RS256). The frontend gets a fresh one with Clerk's `getToken()`
        before each request. The backend verifies signature (Clerk JWKS), exp/nbf, iss and azp. See CONTRACT.md §2.
  parameters:
...
    RideId:
      name: ride_id
      in: path
      required: true
      schema: { type: integer }
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
...
  responses:
    Unauthorized:
      description: Missing, invalid or expired Clerk session token (UNAUTHENTICATED)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated but not allowed (FORBIDDEN, ONBOARDING_REQUIRED, ACCOUNT_DEACTIVATED, NOT_A_PARTICIPANT)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource does not exist (or is not visible to the caller)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Business-rule / state conflict — see `code`
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationError:
      description: Request body/params failed validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable code (see CONTRACT.md §5)
          examples: [RIDE_NOT_OPEN]
        message:
          type: string
          description: Human-readable, not for parsing
        details:
          description: |
            For VALIDATION_ERROR — list of field errors, for display. `field` names the most
            specific location the server can attribute the failure to, and a later version may
            make it *more* specific without that being breaking: branch on `code`, use `field`
            to point at an input, and keep a fallback for a `field` you can't place onto one.
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string, examples: [body.departure_time] }
              message: { type: string }
    Latitude:
      type: number
      minimum: -90
      maximum: 90
    Longitude:
      type: number
      minimum: -180
      maximum: 180
    Money:
      type: string
      pattern: '^\d{1,8}(\.\d{1,2})?$'
      description: |
        Decimal as string, e.g. "25.50" (never a float). Being a `string` is the point: a JSON
        number in a request body is 422 VALIDATION_ERROR, whether or not it would round-trip.
        1 or 2 decimal places on the way in, or none; always 2 on the way out.
      examples: ["25.50"]
...
    RideStatus:
      type: string
      enum: [upcoming, full, in_progress, completed, cancelled]
    RequestStatus:
      type: string
      enum: [pending, approved, rejected, cancelled]
...
    VehiclePublic:
      type: object
      description: Vehicle without the plate — what anyone may see
      required: [make, model, color]
      properties:
        make: { type: string }
        model: { type: string }
        color: { type: string }
...
    UserPublic:
      type: object
      description: What other users may see. No email/phone/DOB.
      required: [id, name, driver_rating_count, passenger_rating_count, created_at]
      properties:
        id: { type: integer }
        name: { type: string }
        driver_rating: { type: [number, "null"] }
        driver_rating_count: { type: integer }
        passenger_rating: { type: [number, "null"] }
        passenger_rating_count: { type: integer }
        vehicle:
          oneOf:
            - { $ref: "#/components/schemas/VehiclePublic" }
            - { type: "null" }
        created_at: { type: string, format: date-time, description: "\"Member since\"" }
...
    RidePreferences:
      type: object
      properties:
        smoking: { type: boolean, default: false, description: "true = smoking allowed" }
        pets: { type: boolean, default: false, description: "true = pets allowed" }
        music: { type: boolean, default: true }
        gender_only:
          type: boolean
          default: false
          description: true = only passengers of the driver's gender (hard filter in matching)
    RidePreferencesPatch:
      type: object
      description: |
        Write shape for RidePreferences, used by POST /rides and PATCH /rides/{ride_id}. No
        defaults: on create the server fills them, and on PATCH an absent key is left alone —
        a shallow merge of all four keys, exactly like user preferences (CONTRACT.md §4 Rides).
      properties:
        smoking: { type: boolean, description: "true = smoking allowed" }
        pets: { type: boolean, description: "true = pets allowed" }
        music: { type: boolean }
        gender_only:
          type: boolean
          description: true = only passengers of the driver's gender (hard filter in matching)
    RideCreate:
      type: object
      required:
        [start_lat, start_lng, start_address, end_lat, end_lng, end_address, departure_time, capacity, price_per_seat]
      properties:
        start_lat: { $ref: "#/components/schemas/Latitude" }
        start_lng: { $ref: "#/components/schemas/Longitude" }
        start_address: { type: string, minLength: 1, maxLength: 255 }
        end_lat: { $ref: "#/components/schemas/Latitude" }
        end_lng: { $ref: "#/components/schemas/Longitude" }
        end_address: { type: string, minLength: 1, maxLength: 255 }
        departure_time:
          type: string
          format: date-time
          description: Must be in the future (422 DEPARTURE_IN_PAST)
        capacity: { type: integer, minimum: 1, maximum: 8 }
        price_per_seat: { $ref: "#/components/schemas/Money" }
        preferences: { $ref: "#/components/schemas/RidePreferencesPatch" }
        notes: { type: [string, "null"], maxLength: 1000 }
    RideUpdate:
      type: object
      description: All fields optional; see PATCH /rides/{id} for restrictions
      properties:
        start_lat: { $ref: "#/components/schemas/Latitude" }
        start_lng: { $ref: "#/components/schemas/Longitude" }
        start_address: { type: string, minLength: 1, maxLength: 255 }
        end_lat: { $ref: "#/components/schemas/Latitude" }
        end_lng: { $ref: "#/components/schemas/Longitude" }
        end_address: { type: string, minLength: 1, maxLength: 255 }
        departure_time: { type: string, format: date-time }
        capacity: { type: integer, minimum: 1, maximum: 8 }
        price_per_seat: { $ref: "#/components/schemas/Money" }
        preferences: { $ref: "#/components/schemas/RidePreferencesPatch" }
        notes: { type: [string, "null"], maxLength: 1000 }
    MyRequest:
      type: object
      description: |
        The caller's own request on a ride, as carried by Ride.my_request.
      required: [id, status, seats_requested]
      properties:
        id: { type: integer, description: "The id to use for POST /requests/{request_id}/cancel" }
        status: { $ref: "#/components/schemas/RequestStatus" }
        seats_requested: { type: integer }
    Ride:
      type: object
      required:
        - id
        - driver
        - start_lat
        - start_lng
        - start_address
        - end_lat
        - end_lng
        - end_address
        - departure_time
        - capacity
        - available_seats
        - price_per_seat
        - status
        - preferences
        - created_at
        - updated_at
      properties:
        id: { type: integer }
        driver: { $ref: "#/components/schemas/UserPublic" }
        start_lat: { $ref: "#/components/schemas/Latitude" }
        start_lng: { $ref: "#/components/schemas/Longitude" }
        start_address: { type: string }
        end_lat: { $ref: "#/components/schemas/Latitude" }
        end_lng: { $ref: "#/components/schemas/Longitude" }
        end_address: { type: string }
        departure_time: { type: string, format: date-time }
        capacity: { type: integer }
        available_seats: { type: integer, minimum: 0 }
        price_per_seat: { $ref: "#/components/schemas/Money" }
        status: { $ref: "#/components/schemas/RideStatus" }
        preferences: { $ref: "#/components/schemas/RidePreferences" }
        notes: { type: [string, "null"] }
        driver_vehicle_plate:
          type: [string, "null"]
          description: Only filled for the driver and for passengers with an approved request on this ride; null for everyone else
        my_request:
          description: |
            The caller's own blocking request on this ride - the single pending/approved/rejected
            row allowed per (ride, passenger). Null when the caller never requested it, after
            their own cancel (so the request button comes back - CONTRACT.md D2), and for a
            driver on their own ride. Inside a RideRequest this is the caller's *current* request
            on that ride, which may differ from the row you are holding: in request lists read
            the outer object, which is strictly more informative.
          oneOf:
            - { $ref: "#/components/schemas/MyRequest" }
            - { type: "null" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    # ── search ──
...